chore: Remove command-line application code to focus "TensorInfo" as a library.
This commit eliminates the "apps" directory, which contained the source code
for the `lstensors` command-line application. The decision was driven by the
desire to concentrate this project solely on providing a robust library.

The evolving command-line tools, including `ckshow`, are now being developed
in the separate "CheckpointTools" repository:
  * https://github.com/martin-rizzo/CheckpointTools

By removing these applications from TensorInfo, the project is streamlined as
an exclusive library, offering developers the core functionalities needed for
tensor info and checkpoint metadata handling.